refactor(Member): add Email class
